microservices/user_service/requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-tools==1.56.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
protobuf==4.23.3
python-dotenv==1.0.0
SQLAlchemy==2.0.17
typing_extensions==4.7.0
Werkzeug==2.3.6
WTForms==3.0.1
